requirements.txt
# Streamlit app
streamlit==1.26.0

# Data processing
pandas==2.2.3
numpy==1.26.0
scikit-learn==1.4.2

# NLP & ML
spacy==3.6.0
nltk==3.8.1
keybert==0.8.0
sentence-transformers==2.2.2


# Resume file handling
python-docx==0.8.11
pdfminer.six==20221105
PyMuPDF==1.22.5

# Optional: For nice tables or audio tagging
tabulate==0.9.0
eyed3==0.9.8

plotly